+++ CHANGELOG.md
@@ -38,6 +38,7 @@ We haven't had the best track record of code/feature changes before this date, b
   <summary>
    Unreleased changes that have landed in master. Click to see more.
   </summary>
+  - [fix] Fix FlowFixMes in FbtTable.access
 </details>
 
 - 0.14.0: